requirements.txt
piexif==1.1.3
Pillow==9.4.0